Permit saving 0 ttl value in non-live; force min 60 ttl in live
Previously, this conditional would mean a ttl of 0 couldn't be saved to
the database. The conditional also didn't reflect the logic we have
elsewhere forcing a minimum of 60 for ttl in live.

Now, the conditional lets us save a ttl of 0 in non-live environments,
and enforces a minimum of 60 saved to the database in live.