Amended WordPress 5.4.2 release.
The WordPress 5.4.2 release included an additional change to the way Pantheon manages protected web paths by adding the list explicitly to the pantheon.upstream.yml file. The paths listed in the new location were missing the trailing '/' character, which changed the behavior of the path matching. This amended release restores the missing trailing slashes. No changes to WordPress itself are included.